move fastL2GovernanceExecute to scenario/utils
